Refactor form submit handler
