+++ common/Auth/Form/index.tsx
@@ -1,12 +1,6 @@
-import {
-  FieldValues,
-  SubmitHandler,
-  UseFormHandleSubmit,
-} from "react-hook-form";
+import { SubmitHandler, UseFormHandleSubmit, useForm } from "react-hook-form";
 import styles from "./index.module.css";
 import { ReactNode } from "react";
-import { postSignup } from "@/api";
-import { useRouter } from "next/router";
 
 export interface FormValues {
   email: string;
@@ -16,27 +10,19 @@ export interface FormValues {
 
 export interface AuthFormProps {
   children?: ReactNode;
-  onSubmit: UseFormHandleSubmit<FormValues>;
+  onSubmit: SubmitHandler<FormValues>;
+  handleSubmit: UseFormHandleSubmit<FormValues, undefined>;
   purpose: string;
 }
 
-function AuthForm({ children, onSubmit, purpose }: AuthFormProps) {
-  const router = useRouter();
-
-  const handleSubmit: SubmitHandler<FormValues> = async (data) => {
-    const { email, password } = data;
-
-    try {
-      await postSignup({ email, password });
-
-      router.push("/folder");
-    } catch (error) {
-      console.error(error);
-    }
-  };
-
+function AuthForm({
+  children,
+  onSubmit,
+  handleSubmit,
+  purpose,
+}: AuthFormProps) {
   return (
-    <form className={styles.formContainer} onSubmit={onSubmit(handleSubmit)}>
+    <form className={styles.formContainer} onSubmit={handleSubmit(onSubmit)}>
       {children}
       <button className={styles.submitBtn} type="submit">
         {(purpose === "signin" && "로그인") ||
